PdkInstallations. Add Decimal as a valid parameter type, and numeric field for Prefixed
