+++ hdl21/pdk/sample_pdk/test_sample_pdk.py
@@ -14,7 +14,7 @@ def test_default():
     assert h.pdk.default() is sample_pdk
 
 
-def gethasmos():
+def mosmodule():
     # Create a simple Module with each of the default-param Mos types
     hasmos = h.Module(name="hasmos")
     z = hasmos.z = h.Signal()
@@ -31,7 +31,7 @@ def gethasmos():
 
 
 def test_compile():
-    hasmos = gethasmos()
+    hasmos = mosmodule()
 
     # Compile it for the PDK
     pkg = h.to_proto(hasmos)
@@ -51,7 +51,7 @@ def test_compile():
 
 
 def test_netlist():
-    hasmos = gethasmos()
+    hasmos = mosmodule()
 
     # Netlist it for the PDK
     pkg = h.to_proto(hasmos)
