+++ pdks/Sky130/sky130/__init__.py
@@ -2,7 +2,7 @@ from hdl21.pdk import register
 
 # Grab our primary PDK-definition module
 from . import sky130
-from .sky130 import modules, Sky130MosParams, Sky130Walker
+from .sky130 import modules, Sky130MosParams, Sky130Walker, Install
 
 # And register as a PDK module
 register(sky130)
